Added structures like objects. Added the sign wich display text when you press up. Player can now attack. Ladder size has changed so player can go down easier. Added some parallax scrolling background and tried to change its color to green. Changed the ground sprite for the special grounds

+++ core/src/paulogaspar/hero/maps/Background.java
@@ -15,7 +15,7 @@ public class Background{
 	private float scale;
 	private float speed;
 	
-	public Background(TextureRegion region,int x,int y,float scale,float speed){
+	public Background(TextureRegion region,float x,float y,float scale,float speed){
 		this.region = region;
 		position = new float[2];
 		position[0] = x;
